Change return type Interface Expression int to double & Implement SpecialVariables

+++ src/main/java/Grammar/Expression/Expression.java
@@ -6,7 +6,7 @@ import GamePlay.Territory;
 import java.util.HashMap;
 
 public interface Expression {
-    int eval(HashMap<String, Integer> bindings, Player p, Territory t) throws EvalError;
+    double eval(HashMap<String, Integer> bindings, Player p, Territory t) throws EvalError;
     void prettyPrint(StringBuilder s);
 }
